add searchkit hook + remove apollo dep from searchkit/client
