+++ packages/searchkit-sdk/package.json
@@ -42,6 +42,7 @@
     "@types/node": "^13.9.1",
     "cross-fetch": "^3.1.5",
     "nock": "^13.0.5",
-    "typescript": "^4.0.5"
+    "typescript": "^4.0.5",
+    "react": "^17.0.2"
   }
 }
